Finalize first version

+++ app/src/main/java/com/dionpapas/inventoryapp/MainActivity.java
@@ -4,6 +4,7 @@ import android.content.Context;
 import android.content.Intent;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
+import android.graphics.Color;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
@@ -16,6 +17,7 @@ import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
+import android.widget.TextView;
 import android.widget.Toast;
 
 import com.dionpapas.inventoryapp.data.InventoryAppContract;
@@ -31,7 +33,9 @@ import java.util.Date;
 import static com.dionpapas.inventoryapp.data.InventoryAppContract.PositionEntry.COLUMN_DIFFERENCE;
 import static com.dionpapas.inventoryapp.data.InventoryAppContract.PositionEntry.COLUMN_ITEM;
 import static com.dionpapas.inventoryapp.data.InventoryAppContract.PositionEntry.COLUMN_POSITION;
+import static com.dionpapas.inventoryapp.data.InventoryAppContract.PositionEntry.COLUMN_STOCK;
 import static com.dionpapas.inventoryapp.data.InventoryAppContract.PositionEntry.COLUMN_TIMESTAMP;
+import static com.dionpapas.inventoryapp.data.InventoryAppContract.PositionEntry.COLUMN_WMS;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -42,8 +46,13 @@ public class MainActivity extends AppCompatActivity {
     private static final String SHARED_PROVIDER_AUTHORITY = BuildConfig.APPLICATION_ID + ".myfileprovider";
     private static final String SHARED_FOLDER = "shared";
     private String to[] = {"theodoropoulosvas@gmail.com"};
+   // private String to[] = {"nionios250@gmail.com"};
     private String emailBody= "Here comes the database export";
     private String emailSubject;
+    private boolean mFileExists;
+    private String filename = "export.csv";
+    private boolean clicked = false;
+    private String orderBy = " ASC";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -51,23 +60,18 @@ public class MainActivity extends AppCompatActivity {
         setContentView(R.layout.activity_main);
         mContext = getApplicationContext();
         RecyclerView waitlistRecyclerView;
-
         waitlistRecyclerView = (RecyclerView) this.findViewById(R.id.all_guests_list_view);
-
         waitlistRecyclerView.setLayoutManager(new LinearLayoutManager(this));
-
         InventoryDBHelper dbHelper = new InventoryDBHelper(this);
-
         mDb = dbHelper.getWritableDatabase();
         //Insert fake data
         //TestUtils.insertFakeData(mDb);
         //Get all data
-        Cursor cursor = getAllRegistrations();
+        Cursor cursor = getAllRegistrations(orderBy);
         // Create an adapter for that cursor to display the data
         mAdapter = new PositionsListAdapter(this, cursor);
         // Link the adapter to the RecyclerView
         waitlistRecyclerView.setAdapter(mAdapter);
-
         new ItemTouchHelper(new ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT | ItemTouchHelper.RIGHT) {
 
             @Override
@@ -83,7 +87,7 @@ public class MainActivity extends AppCompatActivity {
                 //remove from DB
                 removePosition(id);
                 //update the list
-                mAdapter.swapCursor(getAllRegistrations());
+                mAdapter.swapCursor(getAllRegistrations(orderBy));
             }
 
         }).attachToRecyclerView(waitlistRecyclerView);
@@ -99,19 +103,37 @@ public class MainActivity extends AppCompatActivity {
             }
         });
 
+        final TextView mdateTextView = (TextView) this.findViewById(R.id.date_text);
 
+        mdateTextView.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Log.d(LOG_TAG, "Clicking the textView");
+                if(!clicked){
+                    mdateTextView.setBackgroundColor(Color.parseColor("#e5e5e5")); // silver
+                    clicked = true;
+                    orderBy = " DESC";
+                    mAdapter.swapCursor(getAllRegistrations(orderBy));
+                }else {
+                    mdateTextView.setBackgroundColor(Color.parseColor("#80FFFFFF")); // white
+                    clicked = false;
+                    orderBy = " ASC";
+                    mAdapter.swapCursor(getAllRegistrations(orderBy));
+                }
+            }
+        });
     }
 
     @Override
     protected void onResume() {
         super.onResume();
         // re-queries for all tasks
-        mAdapter.swapCursor(getAllRegistrations());
+        mAdapter.swapCursor(getAllRegistrations(orderBy));
         //getSupportLoaderManager().restartLoader(TASK_LOADER_ID, null, this);
     }
 
 
-    private Cursor getAllRegistrations() {
+    private Cursor getAllRegistrations(String orderBy) {
         return mDb.query(
                 InventoryAppContract.PositionEntry.TABLE_NAME_REGISTRATIONS,
                 null,
@@ -119,7 +141,7 @@ public class MainActivity extends AppCompatActivity {
                 null,
                 null,
                 null,
-                COLUMN_TIMESTAMP
+                COLUMN_TIMESTAMP + orderBy
         );
     }
 
@@ -151,7 +173,8 @@ public class MainActivity extends AppCompatActivity {
     public void exportDB() throws IOException {
         SimpleDateFormat sdf = new SimpleDateFormat("dd_MM_yyyy");
         String currentDate = sdf.format(new Date());
-        String filename = "export_"+currentDate.toString() +".csv";
+        //String filename = "export_"+currentDate.toString() +".csv";
+
         emailSubject = "Database export on " + currentDate.toString();
         File shared = createFile(filename);
         writeFile(shared);
@@ -162,75 +185,54 @@ public class MainActivity extends AppCompatActivity {
         final File sharedFolder = new File(getFilesDir(), SHARED_FOLDER);
         sharedFolder.mkdirs();
         File sharedFile = new File(getExternalFilesDir (SHARED_FOLDER), filename);
-        if(!sharedFile.exists())
+        mFileExists = sharedFile.exists();
+        if(!mFileExists) {
             sharedFile.createNewFile();
-            String columnString =   "\"Position\",\"Item\",\"Time\",\"Difference\"";
+            String columnString = "\"Position\",\"Item\",\"Stock\",\"WMS\",\"Difference\",\"Stock\"" + "\n";
             FileOutputStream fos = new FileOutputStream(sharedFile);
             fos.write(columnString.getBytes());
             fos.close();
+            Log.d(LOG_TAG, "Creating file with colums");
+        } else {
+            Log.d(LOG_TAG, "File already exists");
+        }
         return sharedFile;
     }
 
     private void writeFile(File destination ) {
- //       String columnString =   "\"Position\",\"Item\",\"Time\",\"Difference\"";
-//        String combinedString = columnString + "\n" + dataString;
-
-//        public static final String COLUMN_POSITION = "position_name";
-//        public static final String COLUMN_ITEM = "item";
-//        public static final String COLUMN_STOCK = "stock";
-//        public static final String COLUMN_WMS = "wms";
-//        public static final String COLUMN_DIFFERENCE= "difference";
-//        public static final String COLUMN_TIMESTAMP = "timestamp";
-
+        FileOutputStream fos;
         try {
-            FileOutputStream fos = new FileOutputStream(destination);
-            Cursor cursor = getAllRegistrations();
+            fos = new FileOutputStream(destination,true);
+            Cursor cursor = getAllRegistrations(orderBy);
             if (cursor.moveToFirst()) {
                 while (!cursor.isAfterLast()) {
                     String position = cursor.getString(cursor.getColumnIndex(COLUMN_POSITION)) + ",";
                     String item = cursor.getString(cursor.getColumnIndex(COLUMN_ITEM)) + ",";
-                    String timestamp = cursor.getString(cursor.getColumnIndex(COLUMN_TIMESTAMP)) + ",";
-                    String difference = cursor.getString(cursor.getColumnIndex(COLUMN_DIFFERENCE))+ "\n";
-                    String combinedString = position + item + timestamp + difference;
-
+                    String stock = cursor.getString(cursor.getColumnIndex(COLUMN_STOCK)) + ",";
+                    String WMS = cursor.getString(cursor.getColumnIndex(COLUMN_WMS)) + ",";
+                    String difference = cursor.getString(cursor.getColumnIndex(COLUMN_DIFFERENCE))+ ",";
+                    String timestamp = cursor.getString(cursor.getColumnIndex(COLUMN_TIMESTAMP)) + "\n";
+                    String combinedString = position + item + stock + WMS + difference + timestamp;
                     fos.write(combinedString.getBytes());
+                    long id = cursor.getLong(cursor.getColumnIndex(InventoryAppContract.PositionEntry._ID));
+                    removePosition(id);
                     cursor.moveToNext();
                 }
+                fos.close();
             }
-//            outputStream = openFileOutput(destination.getName(), Context.MODE_PRIVATE);
-            fos.close();
-            Toast.makeText(getBaseContext(),"Data saved to file",Toast.LENGTH_SHORT).show();
-            send_email(destination);
         } catch (IOException e) {
             e.printStackTrace();
         }
+//      outputStream = openFileOutput(destination.getName(), Context.MODE_PRIVATE);
+        Toast.makeText(getBaseContext(),"Exporting data to file, cleaning db",Toast.LENGTH_SHORT).show();
+        send_email(destination);
     }
-
-//    public void readFile(File sharedFile){
-//        FileInputStream in = null;
-//        try {
-//            in = openFileInput(sharedFile.getName());
-//        } catch (FileNotFoundException e) {
-//            e.printStackTrace();
-//        }
-//        InputStreamReader inputStreamReader = new InputStreamReader(in);
-//        BufferedReader bufferedReader = new BufferedReader(inputStreamReader);
-//        StringBuilder sb = new StringBuilder();
-//        String line;
-//        try {
-//            while ((line = bufferedReader.readLine()) != null) {
-//                sb.append(line);
-//                Log.d(LOG_TAG,"Reading file to see if it empty" + line);
-//            }
-//        } catch (IOException e) {
-//            e.printStackTrace();
-//        }
-//    }
 
     public void send_email(File file){
         Uri path = null;
         try {
             path = FileProvider.getUriForFile(mContext, SHARED_PROVIDER_AUTHORITY, file );
+
         } catch (IllegalArgumentException e) {
             Log.e("File Selector", "The selected file can't be shared: " + file.getName() + e);
         }
